code/requirements.txt
absl-py==2.0.0
addict==2.4.0
annotated-types==0.6.0
audioread==3.0.1
av==11.0.0
fastapi==0.104.1
gradio==4.7.1
gradio_client==0.7.0
grpcio==1.60.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-resources==6.1.1
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kaggle==1.6.12
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
msgpack==1.0.8
nest-asyncio==1.5.8
nltk==3.8.1
numba==0.59.1
numpy
oauthlib==3.2.2
onnx==1.15.0
onnx-simplifier==0.4.35
opencv-python==4.8.1.78
orjson==3.9.10
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
Pillow 
platformdirs==4.1.0
pooch==1.8.1
proglog==0.1.10
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
python-slugify==8.0.4
pytube==15.0.0
pytz==2023.3.post1
PyYAML 
pyzmq==25.1.2
referencing==0.31.1
regex==2023.10.3
requests
requests-oauthlib==1.3.1
rich==13.7.0
rouge==1.0.1
rpds-py==0.13.2
rsa==4.9
safetensors==0.4.1
scikit-learn==1.4.2
scipy==1.11.4
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
stack-data==0.6.3
starlette==0.27.0
streamlit==1.33.0
tokenizers==0.15.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.8.0
uvicorn==0.24.0.post1
watchdog==4.0.0
webencodings==0.5.1
websockets==11.0.3
